Command Package: Add Assertions
Class files in Command package are not asserted making
them functions in the class files more prone to errors.

Add assertion to functions in command package to assert
functions' behaviour is working as intended.

Assertion has been added to the start of the functions to
assert that the inputs are not null and are of expected values.