Refactor CommandHandler and MessageHandler
Merging master to A-Streams did not merge the Assert Branch properly due
to a previous marge mistake. This commit aims to add the missing assert
statements from the assert branch.

The change is done to add the assert statements and help make the code
more bug free by asserting possible points of the code that can give bugs.

What's done about it is that the assert statements from the Assert Branch has been
added back in to the `CommandHandler.java` and `MessageHandler.java`
by trying to reverse the wrong merge commit.

As for why it's done this way, the reversal of commit helps with
reducing redundant work and just let software like GitHub Desktop
handle changes by clicking on changes to be kept