Add multiline messaging UI feature
Messages sent by Hokmah used to be only one full dialog box for
the UI.

With this commit, Hokmah now has the ability to send messages
through multiple dialog box by making the MessageHandler return a
list of message lines instead of a full string.

Why it was done this way was to add more customizability and flavour
when showing messages through the UI.

Let's
- Change the `hokmah.command` and `hokmah.exception` packages
to give responses as array of String rather than a full String
- Change `MainWindowController.java` to show responses as multiple
dialog boxes instead of just one full one.